+++ components/admin/important/ImportantList.tsx
@@ -32,7 +32,7 @@ export default function ImportantList({
   };
 
   return (
-    <div className="mt-6 mb-8 mx-2.5">
+    <div className="mb-8 mx-2.5">
       <ImportantListHeader />
       <ul className={`divide-y divide-neutral-200 divide-dashed border-b border-neutral-300`}>
         {posts.map((post, i) => (
