+++ components/admin/slide/SlideListRow.tsx
